% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ex_data}
\alias{ex_data}
\title{Access example data layers included in the package}
\usage{
ex_data(
  layer = c("elev_ned_30m.tif", "geology.shp", "lakes.shp", "landuse_r.tif",
    "pointsources.shp", "sites_nc.shp", "streams.shp")
)
}
\arguments{
\item{layer}{Character. Name of the example layer to load. Must be one of:
"elev_ned_30m.tif", "geology.shp", "lakes.shp", "landuse_r.tif",
"pointsources.shp", "sites_nc.shp", "streams.shp"}
}
\value{
A \code{SpatRaster} or \code{SpatVector} object, depending on the layer.
}
\description{
Access example data layers included in the package
}
\details{
Original source of data: https://github.com/MiKatt/openSTARS
}
\examples{
\dontrun{
ex_data("elev_ned_30m.tif") # example DEM raster
ex_data("geology.shp") # example geology vector
ex_data("lakes.shp") # example lakes vector
ex_data("landuse_r.tif") # example land use raster
ex_data("pointsources.shp") # example point sources vector
ex_data("sites_nc.shp") # example sampling sites vector
ex_data("streams.shp") # example stream network vector
}

}
