% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihydro-package.R
\docType{package}
\name{ihydro-package}
\alias{ihydro}
\alias{ihydro-package}
\title{ihydro: Integrated hydrology tools for aquatic environmental science}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implements workflows associated with geospatial analysis of aquatic systems. Uses user defined Digital Elevation Models (DEM) to produce attributed stream line vectors, catchment and subcatchment polygons, and pairwise distances among stream segments. Sampling points may be inserted to stream network.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/p-schaefer/ihydro}
  \item Report bugs at \url{https://github.com/p-schaefer/ihydro/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schaefer \email{ecopulseanalytics@gmail.com}

}
\keyword{internal}
