% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihydro_to_SSN.R
\name{ihydro_to_ssn}
\alias{ihydro_to_ssn}
\title{Convert iHydro stream data to SSN format}
\usage{
ihydro_to_ssn(
  ihydro_obj,
  obs = NULL,
  pred = NULL,
  lsn_path = NULL,
  verbose = FALSE
)
}
\arguments{
\item{ihydro_obj}{An iHydro object}

\item{obs}{A data frame of observed site data with a link_id column}

\item{pred}{A data frame of predicted site data with a link_id column}

\item{lsn_path}{Pathname to a directory in character format specifying where to store the outputs. The directory will be created if it does not already exist..}

\item{verbose}{Logical.}
}
\value{
An SSN object containing the stream network and site data
}
\description{
This function converts stream network data from an iHydro object into a format
compatible with the Spatial Stream Network (SSN) modeling framework. It processes
the stream lines, prepares site data for observed and predicted locations, and assembles
the SSN object with the appropriate attributes for edges and sites. The function also
handles topology errors that may arise during the conversion process.
}
