% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_flowdir.R
\name{process_flowdir}
\alias{process_flowdir}
\title{Process flow direction/accumulation and extract streams from DEM}
\usage{
process_flowdir(
  dem,
  threshold,
  burn_streams = NULL,
  stream_keep_thresh = NULL,
  chunk_size = 250,
  stream_keep_burnbuff = 25,
  stream_keep_dembuff = 1,
  burn_depth = NULL,
  burn_slope_dist = NULL,
  burn_slope_depth = NULL,
  min_length = NULL,
  depression_corr = c(NULL, "fill", "breach"),
  output_filename,
  return_products = FALSE,
  temp_dir = NULL,
  compress = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dem}{Character path, \code{SpatRaster}, or \code{PackedSpatRaster} for a DEM.}

\item{threshold}{Integer. Flow accumulation threshold for stream initiation.}

\item{burn_streams}{Optional stream layer to burn into the DEM.}

\item{stream_keep_thresh}{Numeric minimum proportion of overlap between \code{burn_streams}
and stream segments extracted from DEM to keep in final stream layer (see details)}

\item{burn_depth}{Numeric burn depth in map elevation units (default \code{NULL}).}

\item{burn_slope_dist}{Numeric burn slope width in meters (default \code{NULL}).}

\item{burn_slope_depth}{Numeric burn slope depth in map elevation units (default \code{NULL}).}

\item{min_length}{Numeric. Minimum tributary length in meters; shorter 1st-order
tributaries are removed.}

\item{depression_corr}{\code{NULL}, \code{"fill"} \code{\link[whitebox:wbt_fill_depressions]{whitebox::wbt_fill_depressions()}}, or \code{"breach"} \code{\link[whitebox:wbt_breach_depressions]{whitebox::wbt_breach_depressions()}}.}

\item{output_filename}{Character path ending in \code{.gpkg}.}

\item{return_products}{Logical. Return raster products in addition to the
file path?}

\item{temp_dir}{Character directory for temporary files.}

\item{compress}{Logical. Compress output rasters?}

\item{verbose}{Logical.}
}
\value{
An object of class \code{ihydro}.
The contained rasters are:
\itemize{
\item \code{dem_raw}: the input DEM.
\item \code{dem_final}: the final DEM after burning and depression correction.
\item \code{dem_d8}: D8 flow direction pointer raster.
\item \code{dem_accum_d8}: D8 flow accumulation in number of upstream cells.
\item \code{dem_accum_d8_sca}: D8 flow accumulation in upstream contributing area (SCA).
\item \code{dem_streams_d8}: D8-derived stream raster based on the specified flow accumulation threshold.
\item \code{dem_streams_d8_sub}: Subset of dem_streams_d8 which meet minimum stream_keep_thresh.
If \code{return_products} is \code{TRUE}, the returned \code{ihydro} object will also include the above rasters as \code{SpatRaster} objects for immediate use in R. Otherwise, only file paths are included and temporary files are cleaned up.
}
}
\description{
Produces D8 flow direction, flow accumulation, and stream extraction rasters
from a DEM, storing everything in a GeoPackage.
}
\details{
If \code{burn_streams} is specified:
\itemize{
\item with \code{burn_slope_depth} and \code{burn_slope_dist}, then a gradual slope of
\code{burn_slope_depth} (map depth units) over \code{burn_slope_dist} meters is applied
to directing surface flows to the nearest \code{burn_streams}.
\item with \code{burn_depth}, then the dem covered by the supplied stream lines is depressed
into the landscape by \code{burn_depth} meters.
\item \link[whitebox:wbt_feature_preserving_smoothing]{whitebox::wbt_feature_preserving_smoothing} is applied
\item if \code{stream_keep_thresh} is provided, then the overlap between \code{burn_streams}
and \code{dem_streams_d8} is calculated and only stream segments >= \code{stream_keep_thresh}
will be kept in the final product (note that this option should only be specified
if the \code{burn_streams} is considered highly accurate, and is burned into the DEM
deeply enough that extracted streams closely align)
}

\code{depression_corr} can apply
filling \code{\link[whitebox:wbt_fill_depressions]{whitebox::wbt_fill_depressions()}} or breaching
\code{\link[whitebox:wbt_breach_depressions]{whitebox::wbt_breach_depressions()}} to enforce flow from internal pit cells.
}
\examples{
\dontrun{
# Example usage
tdir <- tempdir()

ex_dem <- ex_data("elev_ned_30m.tif")
ex_streams <- ex_data("streams.shp")
output_gpkg <- file.path(tdir, "output.gpkg")

ihydro_obj <- ihydro::process_flowdir(
   dem = ex_dem,
   threshold = 1000,
   burn_streams = ex_streams,
   burn_depth = 5,
   burn_slope_dist = 250,
   burn_slope_depth = 5,
   min_length = 3,
   depression_corr = "breach",
   output_filename = output_gpkg,
   return_products = TRUE,
   verbose = TRUE
)

ihydro::ihydro_layers(ihydro_obj, n = Inf)

flow_accumulation <- ihydro::read_ihydro(ihydro_obj, "dem_accum_d8")
terra::plot(log10(flow_accumulation), main = "D8 Flow Accumulation")
}

}
