% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_access.R
\name{read_ihydro}
\alias{read_ihydro}
\title{Extract layers stored in an ihydro GeoPackage}
\usage{
read_ihydro(ihydro, layer, collect = TRUE)
}
\arguments{
\item{ihydro}{An object of class \code{ihydro}, as returned by \code{\link[=as.ihydro]{as.ihydro()}},
\code{\link[=process_flowdir]{process_flowdir()}}, or any other ihydro function.}

\item{layer}{Character. Name of the layer to read from the GeoPackage. Use \code{\link[=ihydro_layers]{ihydro_layers()}} to see what layers are available.}

\item{collect}{Logical. Whether to use \code{\link[dplyr:compute]{dplyr::collect()}} to retieve data into a local tibble. Only relevant when \code{layer} is a table.}
}
\value{
The requested layer, read from the GeoPackage. The type of the returned object depends on the layer:
\itemize{
\item Raster layers are returned as \code{SpatRaster} objects.
\item Vector layers are returned as \code{SpatVector} objects.
\item Plain tables are returned as \code{tibble} data frames.
}
}
\description{
Reads a specified layer from the ihydro GeoPackage and returns it as an R object (see return value for details).
}
\examples{
\dontrun{
hydro <- as.ihydro("my_watershed.gpkg")

# List everything
ihydro_layers(hydro)

# Return the stream lines
read_ihydro(hydro, "stream_lines")
}

}
\seealso{
\code{\link[=ihydro_layers]{ihydro_layers()}} to create an \code{ihydro} object.
}
