% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_weights.R
\name{prep_weights}
\alias{prep_weights}
\title{Prepare inverse distance weights}
\usage{
prep_weights(
  input,
  output_filename,
  sample_points = NULL,
  link_id = NULL,
  weighting_scheme = c("iFLS", "HAiFLS", "iFLO", "HAiFLO"),
  inv_function = function(x) (x * 0.001 + 1)^-1,
  return_products = FALSE,
  mem_fraction = 0.5,
  temp_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input}{An \code{ihydro} object.}

\item{output_filename}{Optional path for weight GeoPackage.}

\item{sample_points}{Character vector of site IDs, or \code{NULL} for all.}

\item{link_id}{Character vector of link IDs, or \code{NULL}.}

\item{weighting_scheme}{Character vector of weighting schemes.}

\item{inv_function}{Inverse distance function.}

\item{temp_dir}{Temporary file directory.}

\item{verbose}{Logical.}
}
\value{
An \code{ihydro} object pointing to the weight file.
}
\description{
Computes stream-targeted (iFLS, HAiFLS) and point-targeted (iFLO, HAiFLO)
inverse distance weights and stores them in a GeoPackage.
Stream-targeted weights are calculated for all stream links, while point-targeted
weights are calculated for a subset of target points (e.g., sample points) to save time and storage.
The function checks for existing weights in the output GeoPackage and only calculates missing ones.
It uses parallel processing to speed up the computation of point-targeted weights, which can be time-consuming.
}
\details{
Point-targeted (iFLO, HAiFLO) inverse distance weights are calculated and stored in unnested format,
meaning that weights for different target points may be stored in the same raster layer, if they do
not overlap. This allows for more efficient storage and retrieval of weights for specific target points
without having to load many layers.
}
\examples{
\dontrun{
# Example usage

future::plan(future::multisession(workers = 3))

tdir <- tempdir()

ex_dem <- ex_data("elev_ned_30m.tif")
ex_streams <- ex_data("streams.shp")
ex_points <- ex_data("sites_nc.shp")
output_gpkg <- file.path(tdir, "output.gpkg")

# Process Hydrology
ihydro_obj <- ihydro::process_hydrology(
   dem = ex_dem,
   threshold = 1000,
   burn_streams = ex_streams,
   burn_depth = 5,
   burn_slope_dist = 250,
   burn_slope_depth = 5,
   min_length = 3,
   depression_corr = "breach",
   points = ex_points,
   site_id_col = "site_id",
   snap_distance = 150,
   break_on_noSnap = FALSE,
   pwise_dist = TRUE,
   pwise_all_links = TRUE,
   output_filename = output_gpkg,
   return_products = TRUE,
   verbose = TRUE
)

# Prep Weights
weight_file <- file.path(tdir, "weights.gpkg")
ihydro_weights <- ihydro::prep_weights(
   input = ihydro_obj,
   output_filename = weight_file,
   sample_points = c("1", "25", "80"),
   link_id = c("100", "200", "800"),
   target_o_type = "segment_point",
   weighting_scheme = c("iFLS", "HAiFLS", "iFLO", "HAiFLO"),
   inv_function = function(x) (x * 0.001 + 1)^-1,
   verbose = TRUE
)

lookup_tbl <- ihydro::read_ihydro(ihydro_weights,"unnest_catchment")
weight_id <- dplyr::filter(lookup_tbl, link_id == "800")$unn_group

weight_rast <- ihydro::read_ihydro(ihydro_weights, paste0("iFLO_unn_group", weight_id))

terra::plot(weight_rast)

}

}
