% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hydrology.R
\name{process_hydrology}
\alias{process_hydrology}
\title{Process Hydrology}
\usage{
process_hydrology(
  dem,
  output_filename,
  threshold,
  burn_streams = NULL,
  burn_depth = NULL,
  burn_slope_dist = NULL,
  burn_slope_depth = NULL,
  stream_keep_thresh = NULL,
  chunk_size = 250,
  stream_keep_burnbuff = 25,
  stream_keep_dembuff = 1,
  min_length = NULL,
  depression_corr = c(NULL, "fill", "breach"),
  extra_attr = c("link_slope", "cont_slope", "USChnLn_To", "Elevation", "StOrd_Hack",
    "StOrd_Str", "StOrd_Hort", "StOrd_Shr"),
  points = NULL,
  snap_distance = NULL,
  break_on_noSnap = TRUE,
  site_id_col = NULL,
  pwise_dist = FALSE,
  pwise_all_links = FALSE,
  return_products = FALSE,
  temp_dir = NULL,
  compress = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dem}{Character path, \code{SpatRaster}, or \code{PackedSpatRaster} for a DEM.}

\item{output_filename}{Character path ending in \code{.gpkg}.}

\item{threshold}{Integer. Flow accumulation threshold for stream initiation.}

\item{burn_streams}{Optional stream layer to burn into the DEM.}

\item{burn_depth}{Numeric burn depth in map elevation units (default \code{NULL}).}

\item{burn_slope_dist}{Numeric burn slope width in meters (default \code{NULL}).}

\item{burn_slope_depth}{Numeric burn slope depth in map elevation units (default \code{NULL}).}

\item{stream_keep_thresh}{Numeric minimum proportion of overlap between \code{burn_streams}
and stream segments extracted from DEM to keep in final stream layer (see details)}

\item{min_length}{Numeric. Minimum tributary length in meters; shorter 1st-order
tributaries are removed.}

\item{depression_corr}{\code{NULL}, \code{"fill"} \code{\link[whitebox:wbt_fill_depressions]{whitebox::wbt_fill_depressions()}}, or \code{"breach"} \code{\link[whitebox:wbt_breach_depressions]{whitebox::wbt_breach_depressions()}}.}

\item{extra_attr}{Character vector of optional stream attributes.}

\item{points}{Optional spatial points for sampling locations.}

\item{snap_distance}{Numeric. Max snapping distance in map units.}

\item{break_on_noSnap}{Logical. Error if any point can't be snapped?}

\item{site_id_col}{Character. Column name in \code{points} for site IDs.}

\item{pwise_dist}{Logical. Calculate all downstream flow-connected pairwise distances?}

\item{pwise_all_links}{Logical. Calculate all flow un-connected pairwise distances?}

\item{return_products}{Logical. Return raster products in addition to the
file path?}

\item{temp_dir}{Character directory for temporary files.}

\item{compress}{Logical. Compress output rasters?}

\item{verbose}{Logical.}
}
\value{
An \code{ihydro} object.
}
\description{
This function processes hydrology data by performing flow direction analysis, generating vectors, and tracing flowpaths. It takes a digital elevation model (DEM) and various parameters to customize the processing steps. The function returns an \code{ihydro} object containing the processed hydrology data.
}
\examples{
\dontrun{
# Example usage
tdir <- tempdir()

ex_dem <- ex_data("elev_ned_30m.tif")
ex_streams <- ex_data("streams.shp")
ex_points <- ex_data("sites_nc.shp")
output_gpkg <- file.path(tdir, "output.gpkg")

# Process Hydrology
ihydro_obj <- ihydro::process_hydrology(
   dem = ex_dem,
   threshold = 1000,
   burn_streams = ex_streams,
   burn_depth = 5,
   burn_slope_dist = 250,
   burn_slope_depth = 5,
   min_length = 3,
   depression_corr = "breach",
   points = ex_points,
   site_id_col = "site_id",
   snap_distance = 150,
   break_on_noSnap = FALSE,
   pwise_dist = TRUE,
   pwise_all_links = TRUE,
   output_filename = output_gpkg,
   return_products = TRUE,
   verbose = TRUE
)

}

}
